+++ CHANGELOG.md
@@ -32,6 +32,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Tests for `OrbitResponse`.
 - `UseCase` protocol to define the structure of a use case in an app.
 - Tests for `AgencyResponse`.
+- `GetLaunchListUseCase` for retrieving a list of launches.
 
 ### Changed